requirements.txt
python-dotenv
pydantic
boto3
paramiko